Fixed msg_type issue of the log table.
git-svn-id: http://svn.modules.centreon.com/centreon-broker/trunk/centreon-broker@333 7c7d615a-03b3-410d-a316-1135b1bd9499

+++ inc/events/log.h
@@ -43,7 +43,7 @@ namespace              Events
    public:
     time_t             c_time;
     std::string        host;
-    int                msg_type;
+    short              msg_type;
     std::string        notification_cmd;
     std::string        notification_contact;
     std::string        output;
